Split C function emission
