+++ tests/docs_truth/repo_hygiene/codegen_c.rs
@@ -1,0 +1,29 @@
+use super::*;
+
+#[test]
+fn codegen_c_function_emission_lives_in_focused_helper() {
+    let types = read("src/codegen/c/types.rs");
+    let functions = read("src/codegen/c/functions.rs");
+    let c_mod = read("src/codegen/c/mod.rs");
+
+    for helper in [
+        "emit_function_forward_decl",
+        "emit_function",
+        "format_params",
+        "emit_global",
+    ] {
+        assert!(
+            !types.contains(&format!("fn {helper}")),
+            "C type emission should not own function/global helper: {helper}"
+        );
+        assert!(
+            functions.contains(&format!("fn {helper}")),
+            "C function/global emission should live in focused helper: {helper}"
+        );
+    }
+
+    assert!(
+        c_mod.contains("mod functions;"),
+        "C codegen should load focused function/global emission helper"
+    );
+}
